feat: add tests

+++ ldes-server-retention/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/retention/services/retentionpolicy/execution/MemberRemoverImpl.java
@@ -34,9 +34,11 @@ public class MemberRemoverImpl implements MemberRemover {
 	@Override
 	public void removeMembersFromEventSource(List<MemberProperties> memberProperties) {
 		List<String> ids = memberProperties.stream().filter(MemberProperties::isInEventSource).map(MemberProperties::getId).toList();
-		memberPropertiesRepository.removeFromEventSource(ids);
-		applicationEventPublisher.publishEvent(
-				new MembersRemovedFromEventSourceEvent(ids));
+		if (!ids.isEmpty()) {
+			memberPropertiesRepository.removeFromEventSource(ids);
+			applicationEventPublisher.publishEvent(
+					new MembersRemovedFromEventSourceEvent(ids));
+		}
 	}
 
 	@Override
